ktl-657 refactor: Next.js structure
