+++ blocks/community/community-banner/community-banner.module.css
@@ -1,6 +1,3 @@
-.banner {
-}
-
 .content {
     margin: 24px 0 var(--ktl-grid-gutter);
     max-width: 720px;
