+++ blocks/community/kugs-banner/kugs-banner.tsx
